Test using userdata to boot iPXE ISO
To be able to bootstrap a empty block device in Cloudstack we need to
be able to inject some netboot information to the host. We do this by
adding userdata which the node will download from the DHCP server on
boot. The data will contain the neccessary information to continue the
boot procedure and eventually start the kickstart installation.